+++ lib/ReactViews/Workbench/Controls/WorkbenchItemControls.tsx
@@ -1,10 +1,11 @@
 import { observer } from "mobx-react";
 import { FC } from "react";
+import { isJsonObject } from "../../../Core/Json";
 import TerriaError from "../../../Core/TerriaError";
-import { Complete } from "../../../Core/TypeModifiers";
+import CatalogMemberMixin from "../../../ModelMixins/CatalogMemberMixin";
 import DiscretelyTimeVaryingMixin from "../../../ModelMixins/DiscretelyTimeVaryingMixin";
-import hasTraits from "../../../Models/Definition/hasTraits";
 import { BaseModel } from "../../../Models/Definition/Model";
+import hasTraits from "../../../Models/Definition/hasTraits";
 import {
   DEFAULT_PLACEMENT,
   SelectableDimension
@@ -25,66 +26,45 @@ import DimensionSelectorSection from "./SelectableDimensionSection";
 import ShortReport from "./ShortReport";
 import TimerSection from "./TimerSection";
 import ViewingControls from "./ViewingControls";
-
-type WorkbenchControls = {
-  viewingControls?: boolean;
-  opacity?: boolean;
-  scaleWorkbench?: boolean;
-  splitter?: boolean;
-  timer?: boolean;
-  chartItems?: boolean;
-  filter?: boolean;
-  dateTime?: boolean;
-  timeFilter?: boolean;
-  selectableDimensions?: boolean;
-  colorScaleRange?: boolean;
-  shortReport?: boolean;
-  legend?: boolean;
-};
+import {
+  WorkbenchControls,
+  disableAllControls,
+  enableAllControls
+} from "./WorkbenchControls";
 
 type WorkbenchItemControlsProps = {
   item: BaseModel;
   viewState: ViewState;
-  /** Flag to show each control - defaults to all true */
-  controls?: WorkbenchControls;
-};
+  /**
+   * Disable viewing controls menua
+   */
+  disableViewingControlsMenu?: boolean;
 
-export const defaultControls: Complete<WorkbenchControls> = {
-  viewingControls: true,
-  opacity: true,
-  scaleWorkbench: true,
-  splitter: true,
-  timer: true,
-  chartItems: true,
-  filter: true,
-  dateTime: true,
-  timeFilter: true,
-  selectableDimensions: true,
-  colorScaleRange: true,
-  shortReport: true,
-  legend: true
-};
-
-export const hideAllControls: Complete<WorkbenchControls> = {
-  viewingControls: false,
-  opacity: false,
-  scaleWorkbench: false,
-  splitter: false,
-  timer: false,
-  chartItems: false,
-  filter: false,
-  dateTime: false,
-  timeFilter: false,
-  selectableDimensions: false,
-  colorScaleRange: false,
-  shortReport: false,
-  legend: false
+  /**
+   * Flags to show/hide controls, disableAll=true will disable all controls by default
+   */
+  controls?: Partial<WorkbenchControls>;
 };
 
 const WorkbenchItemControls: FC<WorkbenchItemControlsProps> = observer(
-  ({ item, viewState, controls: controlsWithoutDefaults }) => {
-    // Apply controls from props on top of defaultControls
-    const controls = { ...defaultControls, ...controlsWithoutDefaults };
+  ({
+    item,
+    viewState,
+    controls: propsControls = {},
+    disableViewingControlsMenu
+  }) => {
+    const itemControls =
+      CatalogMemberMixin.isMixedInto(item) &&
+      isJsonObject(item.workbenchControlFlags)
+        ? (item.workbenchControlFlags as Partial<WorkbenchControls>)
+        : undefined;
+
+    // disable/enable all controls, props controls overrides item controls
+    const disableAll = !!(propsControls.disableAll ?? itemControls?.disableAll);
+    const controls = disableAll
+      ? { ...disableAllControls, ...itemControls, ...propsControls, disableAll }
+      : { ...enableAllControls, ...itemControls, ...propsControls, disableAll };
+
     const { generatedControls, error } = generateControls(viewState, item);
 
     if (error) {
@@ -93,22 +73,26 @@ const WorkbenchItemControls: FC<WorkbenchItemControlsProps> = observer(
 
     return (
       <>
-        {controls?.viewingControls ? (
-          <ViewingControls item={item} viewState={viewState} />
-        ) : null}
-        {controls?.opacity ? <OpacitySection item={item} /> : null}
-        {controls?.scaleWorkbench ? <ScaleWorkbenchInfo item={item} /> : null}
-        {controls?.timer ? <TimerSection item={item} /> : null}
-        {controls?.splitter ? <LeftRightSection item={item as any} /> : null}
-        {controls?.chartItems ? <ChartItemSelector item={item} /> : null}
-        {controls?.filter ? <FilterSection item={item} /> : null}
-        {controls?.dateTime && DiscretelyTimeVaryingMixin.isMixedInto(item) ? (
+        {disableViewingControlsMenu ? null : (
+          <ViewingControls
+            item={item}
+            viewState={viewState}
+            controls={controls}
+          />
+        )}
+        {controls.opacity ? <OpacitySection item={item} /> : null}
+        {controls.scaleWorkbench ? <ScaleWorkbenchInfo item={item} /> : null}
+        {controls.timer ? <TimerSection item={item} /> : null}
+        {controls.compare ? <LeftRightSection item={item as any} /> : null}
+        {controls.chartItems ? <ChartItemSelector item={item} /> : null}
+        {controls.filter ? <FilterSection item={item} /> : null}
+        {controls.dateTime && DiscretelyTimeVaryingMixin.isMixedInto(item) ? (
           <DateTimeSelectorSection item={item} />
         ) : null}
-        {controls?.timeFilter ? (
+        {controls.timeFilter ? (
           <SatelliteImageryTimeFilterSection item={item} />
         ) : null}
-        {controls?.selectableDimensions ? (
+        {controls.selectableDimensions ? (
           <DimensionSelectorSection item={item} placement={DEFAULT_PLACEMENT} />
         ) : null}
         {
@@ -120,7 +104,7 @@ const WorkbenchItemControls: FC<WorkbenchItemControlsProps> = observer(
         }
         {/* TODO: remove min max props and move the checks to
       ColorScaleRangeSection to keep this component simple. */}
-        {controls?.colorScaleRange &&
+        {controls.colorScaleRange &&
           hasTraits(
             item,
             WebMapServiceCatalogItemTraits,
@@ -137,9 +121,9 @@ const WorkbenchItemControls: FC<WorkbenchItemControlsProps> = observer(
               maxValue={item.colorScaleMaximum}
             />
           )}
-        {controls?.shortReport ? <ShortReport item={item} /> : null}
-        {controls?.legend ? <Legend item={item} /> : null}
-        {controls?.selectableDimensions ? (
+        {controls.shortReport ? <ShortReport item={item} /> : null}
+        {controls.legend ? <Legend item={item} /> : null}
+        {controls.selectableDimensions ? (
           <DimensionSelectorSection item={item} placement={"belowLegend"} />
         ) : null}
         {